+++ math/bits.v
@@ -4,16 +4,17 @@
 module math
 
 const (
-	uvnan = 0x7FF8000000000001
-	uvinf = 0x7FF0000000000000
-	uvneginf = 0xFFF0000000000000
-	uvone = 0x3FF0000000000000
-	mask = 0x7FF
-	shift = 64 - 11 - 1
-	bias = 1023
-	sign_mask = (u64(1)<<63)
-	frac_mask = ((u64(1)<<u64(shift)) - u64(1))
+	uvnan     = u64(0x7FF8000000000001)
+	uvinf     = u64(0x7FF0000000000000)
+	uvneginf  = u64(0xFFF0000000000000)
+	uvone     = u64(0x3FF0000000000000)
+	mask      = 0x7FF
+	shift     = 64 - 11 - 1
+	bias      = 1023
+	sign_mask = (u64(1) << 63)
+	frac_mask = ((u64(1) << u64(shift)) - u64(1))
 )
+
 // inf returns positive infinity if sign >= 0, negative infinity if sign < 0.
 pub fn inf(sign int) f64 {
 	v := if sign >= 0 { uvinf } else { uvneginf }
@@ -47,15 +48,15 @@ pub fn is_inf(f f64, sign int) bool {
 }
 
 pub fn is_finite(f f64) bool {
-        return !is_nan(f) && !is_inf(f, 0)
+	return !is_nan(f) && !is_inf(f, 0)
 }
 
 // normalize returns a normal number y and exponent exp
 // satisfying x == y × 2**exp. It assumes x is finite and non-zero.
 pub fn normalize(x f64) (f64, int) {
-        smallest_normal := 2.2250738585072014e-308 // 2**-1022
-        if abs(x) < smallest_normal {
-                return x * (1 << 52), -52
-        }
-        return x, 0
+	smallest_normal := 2.2250738585072014e-308 // 2**-1022
+	if abs(x) < smallest_normal {
+		return x * (1 << 52), -52
+	}
+	return x, 0
 }
